requirements.txt
streamlit
google-generativeai
langchain 
PyPDF2
faiss-cpu
langchain_google_genai
python-dotenv